Update file structure and remove a lot of commented code in proc.c
