+++ Makefile
@@ -31,6 +31,7 @@ OBJS = \
   $K/virtio_disk.o \
   $K/buddy.o \
   $K/list.o \
+  $K/proc_list.o \
 
 # riscv64-unknown-elf- or riscv64-linux-gnu-
 # perhaps in /opt/riscv/bin
